SMP/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
braceexpand==0.1.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
discord.py==2.0.1
frozenlist==1.3.1
idna==3.4
import-expression==1.1.4
importlib-metadata==5.0.0
jishaku==2.5.1
line-profiler==3.5.1
multidict==6.0.2
requests==2.28.1
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.12
yarl==1.8.1
zipp==3.8.1
seaborn
matplotlib
numpy